refactor: adding 'Home' folder in assets to organize & change the refs in code.
